Rename walk_pruner to pruner.
